// generated/api-client/models/comment-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { CommentResponseDtoAuthor } from './comment-response-dto-author';

/**
 * 
 * @export
 * @interface CommentResponseDto
 */
export interface CommentResponseDto {
    /**
     * 
     * @type {string}
     * @memberof CommentResponseDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof CommentResponseDto
     */
    'content': string;
    /**
     * 
     * @type {number}
     * @memberof CommentResponseDto
     */
    'likeCount': number;
    /**
     * 
     * @type {number}
     * @memberof CommentResponseDto
     */
    'replyCount': number;
    /**
     * 
     * @type {CommentResponseDtoAuthor}
     * @memberof CommentResponseDto
     */
    'author': CommentResponseDtoAuthor;
    /**
     * 
     * @type {string}
     * @memberof CommentResponseDto
     */
    'parentId': string | null;
    /**
     * 
     * @type {string}
     * @memberof CommentResponseDto
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof CommentResponseDto
     */
    'updatedAt': string;
    /**
     * 
     * @type {boolean}
     * @memberof CommentResponseDto
     */
    'isLiked': boolean | null;
}